'Account' app code refactoring
